Auto complete for path input box.
When user types '/' or '\' in path input box, ajax call will be made to server to query directory listings.  Directory listing api call returns only the directories within given directory, not all files within given directory.  There is no security to what directories are accessible so this might be an upgrade that needs to be done.

Change Synopsis

* added jquery ui 1.10.3
* added auto-complete.js
* server now has an api call that returns directory listing

Possible upgrade

* Directory listing security