Provide git version override check reminder
If failing the check, provide a useful note about how to override that check.
When I'm running in a cygwin environment everything seems to work, but I'm constantly forgetting to set this check.  I then have to look this setting up. Putting in the error msg is more human friendly.

Hope this helps the next person in my situation.